chore(Semaforizacion): Se cambiaron las validaciones al otro modelo
