+++ talent360/evaluaciones/models/evaluacion.py
@@ -2,6 +2,7 @@ from odoo import api, models, fields, _
 from collections import defaultdict, Counter
 from odoo import exceptions
 from datetime import timedelta
+from odoo.exceptions import ValidationError
 
 
 class Evaluacion(models.Model):
@@ -1207,6 +1208,45 @@ class Evaluacion(models.Model):
             "target": "new",
         }
     
+    @api.constrains("niveles")
+    def _check_techo(self):
+        for nivel in self.niveles:
+            # Verificar que el valor de la ponderación sea mayor que 0
+            if nivel.techo <= 0:
+                raise ValidationError(
+                    "El valor de la ponderación debe ser mayor que 0."
+                )
+
+            # Verificar que no haya valores duplicados en la ponderación para la misma evaluación
+            techos = self.niveles.filtered(lambda n: n.id != nivel.id).mapped("techo")
+            if nivel.techo in techos:
+                raise ValidationError(
+                    "No puede haber valores duplicados en la ponderación."
+                )
+
+        todos_techos = self.niveles.mapped("techo")
+        # Verificar que no haya más de 10 techos
+        if len(todos_techos) > 10:
+            raise ValidationError(
+                "No puede haber más de 10 valores de ponderación."
+            )
+        
+        # Verificar que los valores de la ponderación estén en orden ascendente
+        
+        if todos_techos != sorted(todos_techos):
+            raise ValidationError(
+                "Los valores de la ponderación deben estar en orden ascendente."
+            )
+
+        # Verificar que el valor de los 'techos' no sean mayores a 100 y que el último sea 100
+        if todos_techos[-1] > 100:
+            raise ValidationError(
+                "El valor de la ponderación no puede ser mayor a 100."
+            )
+        if todos_techos[-1] != 100:
+            raise ValidationError("El último valor de la ponderación debe ser 100.")
+
+
     def evaluacion_general_action_form(self):
         """
         Ejecuta la acción de redireccionar a la evaluación general y devuelve un diccionario
